requirements.txt
streamlit
tensorflow
Pillow
numpy
pandas
streamlit-drawable-canvas
scikit-learn
matplotlib